remove range.rs
